pyproject.toml
[tool.poetry]
name = "oobabot"
version = "0.3.2"
description = "A Discord bot which talks to Large Language Model AIs using just about any API-enabled backend."
authors = ["Christopher Rude <chris@rudesoftware.net>", "Isabella Nightshade <bella@nightshade.fun>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/xBelladonna/oobabot"

[tool.poetry.dependencies]
aiohttp = "^3.8.4"
python = "^3.8.1"
"discord.py" = "^2.2.2"
pysbd = "^0.3.4"
ruamel-yaml = "^0.18.6"
pynacl = "^1.5.0"
typing-extensions = "^4.9.0"
emoji = "^2.10.1"
pillow = "^10.2.0"

[tool.poetry.scripts]
oobabot = 'oobabot.oobabot:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flakeheaven]
format = "grouped"
max-line-length = 88
show_source = true

[tool.flakeheaven.exceptions]
pycodestyle = ["-E203", "-W503", "-W504"]

[tool.poetry.group.test.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
pre-commit = "^3.2.0"
pyright = "^1.1.325"
pytest = "^7.1"
pylint = "^2.17.4"

[tool.isort]
profile = "black"
force_single_line = true
force_sort_within_sections = true
single_line_exclusions = ["typing", "typing_extensions"]

[tool.pylint.'MAIN']
py-version = '3.8.1'
j = 8

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 88
disable = "C0116,R0902,R0903,R0912,R0913,R0914,W0511,W0621"
include-naming-hint = true
